+++ src/commands/hardis/package/install.ts
@@ -165,63 +165,6 @@ Assisted menu to propose to update \`installedPackages\` property in \`.sfdx-har
       await managePackageConfig(installedPackages, packagesToInstallCompleted);
     }
 
-    /* disabled until sfdx multiple package deployment is working >_<
-        // Post install actions
-        if (!isCI && fs.existsSync(this.sfdxProjectJsonFileName)) {
- 
-            const postInstallResponse = await prompts([
-                {
-                    type: 'confirm',
-                    name: 'retrieve',
-                    message: c.cyanBright('Do you want to retrieve installed package sources in your local branch ?'),
-                    initial: true
-                },
-                {
-                    type: 'confirm',
-                    name: 'sfdxProject',
-                    message: c.cyanBright('Do you want to update your sfdx-project.json ? (advice: yes)'),
-                    initial: true
-                }
-            ]);
-            // Retrieve package sources if requested
-            if (postInstallResponse.retrieve === true) {
-                for (const pckg of packagesToInstallCompleted) {
-                    const retrieveCommand = 'sfdx force:source:retrieve' +
-                    ` -n ${pckg.key}` +
-                    // ` -p ./force-app/main/default` + // let's try without it
-                    ' -w 60';
-                    try {
-                        await execCommand(retrieveCommand, this, { output: true, fail: true, debug: debugMode });
-                    } catch (e) {
-                        // Ugly workaround but it's a sfdx bug...
-                        uxLog(this, c.yellow(`Error while retrieving ${c.bold(pckg.key)} but it may have worked anyway`));
-                        if (fs.existsSync(path.join('.', pckg.key))) {
-                            await fs.remove(path.join('.', pckg.key));
-                        }
-                    }
-                }
-            }
-            // Update sfdx-project.json with new unlocked packages folder references, so it is taken in account with force:source:push and force:source:pull
-            if (postInstallResponse.sfdxProject === true) {
-                const sfdxProjectRaw = await fs.readFile(this.sfdxProjectJsonFileName, 'utf8');
-                const sfdxProject = JSON.parse(sfdxProjectRaw);
-                let updated = false;
-                for (const installedPackage of installedPackages) {
-                    const matchInstalled = packagesToInstallCompleted.filter(pckg => pckg.key === installedPackage.SubscriberPackageName);
-                    const matchLocal = sfdxProject.packageDirectories.filter(packageDirectory => installedPackage.SubscriberPackageName === packageDirectory.path);
-                    if (matchInstalled.length > 0 && matchLocal.length === 0) {
-                        sfdxProject.packageDirectories.push({path: installedPackage.SubscriberPackageName});
-                        updated = true;
-                    }
-                }
-                if (updated) {
-                    await fs.writeFile(this.sfdxProjectJsonFileName, JSON.stringify(sfdxProject, null, 2));
-                    uxLog(this, c.cyan('[config] Updated sfdx-project.json to add new package folders'));
-                }
-            }
-        }
-        */
-
     // Return an object to be displayed with --json
     return { outputString: "Installed package(s)" };
   }