+++ src/commands/hardis/source/push.ts
@@ -1,7 +1,9 @@
 import { FlagsConfig, flags, SfdxCommand } from "@salesforce/command";
 import { Duration } from "@salesforce/kit";
+import * as c from "chalk";
 import { AnyJson } from "@salesforce/ts-types";
 import { wrapSfdxCoreCommand } from "../../../common/utils/wrapUtils";
+import { uxLog } from "../../../common/utils";
 
 export default class Push extends SfdxCommand {
   public static readonly description = `sfdx-hardis wrapper for sfdx force:source:push that displays tips to solve deployment errors.
@@ -40,6 +42,9 @@ export default class Push extends SfdxCommand {
   protected static requiresProject = true;
 
   public async run(): Promise<AnyJson> {
+    uxLog(this,c.red("This command will be deprecated by Salesforce in November 2024."));
+    uxLog(this,c.red("Please migrate to command sf hardis project deploy start"));
+    uxLog(this,c.red("See https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_mig_deploy_retrieve.htm"));
     return await wrapSfdxCoreCommand("sfdx force:source:push", this.argv, this, this.flags.debug);
   }
 }