+++ src/common/utils/deployTipsList.ts
@@ -6,7 +6,7 @@ export function getAllTips() {
       expressionRegex: [/Error (.*) The (.*) apiVersion can't be "([0-9]+)"/gm],
       tip: `{1} metadata has probably been created/updated in a sandbox already upgraded to next platform version (ex: Sandbox in Summer'23 and Production in Spring'23)
 - First, try to update the api version in the XML of {1} metadata file (decrement the number in <apiVersion>{3}.0</apiVersion>)
-- If it still doesn't work because the metadata structure has changed between version, you may try a force:source:retrieve of the metadata by forcing --apiversion at the end of the command.
+- If it still doesn't work because the metadata structure has changed between version, you may try a sf project:retrieve:start of the metadata by forcing --api-version at the end of the command.
       `,
     },
     {
@@ -196,7 +196,7 @@ Example of element to delete:
       label: "Missing e-mail template",
       expressionRegex: [/In field: template - no EmailTemplate named (.*) found/gm],
       tip: `An email template should be present in the sources. To retrieve it, you can run:
-sfdx force:source:retrieve -m EmailTemplate:{1} -u YOUR_ORG_USERNAME`,
+sf project retrieve start -m EmailTemplate:{1} -o YOUR_ORG_USERNAME`,
     },
     {
       name: "empty-item",
@@ -490,7 +490,7 @@ sfdx hardis:project:clean:references , then select "ProductRequest references"`,
       label: "Missing report",
       expressionRegex: [/Error (.*) The (.*) report chart has a problem with the "reportName" field/gm],
       tip: `{1} is referring to unknown report {2}. To retrieve it, you can run:
-- sfdx force:source:retrieve -m Report:{2} -u YOUR_ORG_USERNAME
+- sf project retrieve start -m Report:{2} -o YOUR_ORG_USERNAME
 - If it fails, looks for the report folder and add it before report name to the retrieve command (ex: MYFOLDER/MYREPORTNAME)
 `,
     },
@@ -553,7 +553,7 @@ Go manually make the change in the target org, so the deployment will pass
       label: "Picklist value not found",
       expressionRegex: [/Picklist value: (.*) in picklist: (.*) not found/gm],
       tip: `Sources have references to value {1} of picklist {2}
-- If picklist {2} is standard, add the picklist to sfdx sources by using "sfdx force:source:retrieve -m StandardValueSet:{2}", then save again
+- If picklist {2} is standard, add the picklist to sfdx sources by using "sf project retrieve start -m StandardValueSet:{2}", then save again
 - Else, perform a search in all code of {1}, then remove XML tags referring to {1} (for example in record types metadatas)
 `,
     },
@@ -585,7 +585,7 @@ Go manually make the change in the target org, so the deployment will pass
       label: "CRM Analytics: A Recipe must specify a DataFlow",
       expressionRegex: [/Error (.*) A Recipe must specify a Dataflow/gm],
       tip: `You must include related WaveDataFlow {1} in sources (and probably in package.xml too).
-To retrieve it, run: sfdx force:source:retrieve -m WaveDataFlow:{1} -u SOURCE_ORG_USERNAME
+To retrieve it, run: sf project retrieve start -m WaveDataFlow:{1} -u SOURCE_ORG_USERNAME
 You can also retrieve all analytics sources in one shot using sfdx hardis:org:retrieve:source:analytics -u SOURCE_ORG_USERNAME
   - https://salesforce.stackexchange.com/a/365453/33522
   - https://help.salesforce.com/s/articleView?id=000319274&type=1`,